move CodeCompletions{Client,Params} and CompletionResponseGenerator to lib/shared
