+++ lib/shared/src/index.ts
@@ -107,7 +107,12 @@ export type { GraphContextFetcher } from './graph-context'
 export { GuardrailsPost, summariseAttribution } from './guardrails'
 export type { Attribution, Guardrails } from './guardrails'
 export { SourcegraphGuardrailsClient } from './guardrails/client'
-export { STOP_REASON_STREAMING_CHUNK } from './inferenceClient/misc'
+export {
+    STOP_REASON_STREAMING_CHUNK,
+    type CodeCompletionsClient,
+    type CodeCompletionsParams,
+    type CompletionResponseGenerator,
+} from './inferenceClient/misc'
 export type { IntentClassificationOption, IntentDetector } from './intent-detector'
 export { SourcegraphIntentDetectorClient } from './intent-detector/client'
 export type {
