+++ lib/shared/src/inferenceClient/misc.ts
@@ -1,3 +1,6 @@
+import type { CompletionsClientConfig } from '../sourcegraph-api/completions/client'
+import type { CompletionParameters, CompletionResponse } from '../sourcegraph-api/completions/types'
+
 /**
  * Marks the yielded value as an incomplete response.
  *
@@ -5,3 +8,11 @@
  * all possible response types.
  */
 export const STOP_REASON_STREAMING_CHUNK = 'cody-streaming-chunk'
+
+export type CodeCompletionsParams = Omit<CompletionParameters, 'fast'> & { timeoutMs: number }
+export type CompletionResponseGenerator = AsyncGenerator<CompletionResponse>
+
+export interface CodeCompletionsClient<T = CodeCompletionsParams> {
+    complete(params: T, abortController: AbortController): CompletionResponseGenerator
+    onConfigurationChange(newConfig: CompletionsClientConfig): void
+}
